+++ otelconf/v0.3.0/metric_test.go
@@ -1536,7 +1536,8 @@ func TestPrometheusReaderHostParsing(t *testing.T) {
 			require.NotNil(t, reader)
 
 			t.Cleanup(func() {
-				require.NoError(t, reader.Shutdown(t.Context()))
+				//nolint:usetesting // required to avoid getting a canceled context at cleanup.
+				require.NoError(t, reader.Shutdown(context.Background()))
 			})
 
 			rws, ok := reader.(readerWithServer)
@@ -1568,7 +1569,8 @@ func TestPrometheusReaderConfigurationOptions(t *testing.T) {
 	require.NotNil(t, reader)
 
 	t.Cleanup(func() {
-		require.NoError(t, reader.Shutdown(t.Context()))
+		//nolint:usetesting // required to avoid getting a canceled context at cleanup.
+		require.NoError(t, reader.Shutdown(context.Background()))
 	})
 
 	rws, ok := reader.(readerWithServer)